runtime/pyproject.toml
[tool.poetry]
name = "asam-qc-runtime"
version = "0.1.0"
description = "Python runtime module responsible of executing the qc-framework modules specified in an input configuration file."
authors = ["Danilo Romano <danilo@ivex.ai>","Patrick Abrahão <patrick@ivex.ai>","Tung Dinh <tung@ivex.ai>"]
license = "MPL-2.0"
readme = "../doc/manual/runtime_module.md"
packages = [
    { include = "runtime" },
]

[tool.poetry.dependencies]
python = "^3.10"
asam-qc-baselib = {git = "https://github.com/asam-ev/qc-baselib-py.git", rev = "develop"}
pydantic = "^2.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
black = "^24.4.2"

[tool.poetry.scripts]
qc_runtime = 'runtime.runtime:main'
qc-report-json = 'runtime.report.json.__main__:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
